requirements.txt
# Core dependencies
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3

# Production WSGI server
waitress==3.0.0